minor edits: added a return into checking proxy moved crawler exceptions to new directory
